Shim for XoopsSystemGui class
Changes to make XOOPS 2.5.8 work without errors in strict mode
required a change in the definition of XoopsSystemGui::validate().
These changes will cause errors with old admin themes. This shim
declares the validate() method appropriate to the base class to
allow the admin theme to work on both 2.5.8 and older versions.