Workaround mysql_real_escape_string
Since XOOPS 2.5.7, there is an escape() method in the database
class. In 2.5.8, all use of mysql_* functions was removed from
core, to fix warnings in PHP5, and fatal errors in PHP7. The
mysqli is now used exclusively.

These changes will use the escape() method if it is available,
reverting to mysql_real_escape_string() if needed. This enables
testing on XOOPS 2.5.8 with RMC in PHP7.